Split the loop and handle functions apart to make maintenance easier in the future
